use unsigned literals instead of clumsy casts in ad_parser.c and btstack_event.h

+++ src/ad_parser.c
@@ -63,18 +63,18 @@ void ad_iterator_init(ad_context_t *context, uint8_t ad_len, const uint8_t * ad_
 
 bool ad_iterator_has_more(const ad_context_t * context){
     // assert chunk_len and chunk_type are withing buffer
-    if ((context->offset + (uint8_t) 1) >= context->length) {
+    if ((context->offset + 1u) >= context->length) {
         return false;
     }
 
     // assert chunk_len > 0
     uint8_t chunk_len = context->data[context->offset];
-    if (chunk_len == (uint8_t)  0){
+    if (chunk_len == 0u){
         return false;
     }
     
     // assert complete chunk fits into buffer
-    if ((context->offset + (uint8_t) 1 + chunk_len) > context->length) {
+    if ((context->offset + 1u + chunk_len) > context->length) {
         return false;
     }
     return true;
@@ -83,19 +83,19 @@ bool ad_iterator_has_more(const ad_context_t * context){
 // pre: ad_iterator_has_more() == 1
 void ad_iterator_next(ad_context_t * context){
     uint8_t chunk_len = context->data[context->offset];
-    context->offset += (uint8_t) 1 + chunk_len;
+    context->offset += 1u + chunk_len;
 }
 
 uint8_t   ad_iterator_get_data_len(const ad_context_t * context){
-    return context->data[context->offset] - (uint8_t)  1;
+    return context->data[context->offset] - 1u;
 }
 
 uint8_t   ad_iterator_get_data_type(const ad_context_t * context){
-    return context->data[context->offset + (uint8_t) 1];
+    return context->data[context->offset + 1u];
 }
 
 const uint8_t * ad_iterator_get_data(const ad_context_t * context){
-    return &context->data[context->offset + (uint8_t) 2];
+    return &context->data[context->offset + 2u];
 }
 
 bool ad_data_contains_uuid16(uint8_t ad_len, const uint8_t * ad_data, uint16_t uuid16){
@@ -112,7 +112,7 @@ bool ad_data_contains_uuid16(uint8_t ad_len, const uint8_t * ad_data, uint16_t u
         switch (data_type){
             case BLUETOOTH_DATA_TYPE_INCOMPLETE_LIST_OF_16_BIT_SERVICE_CLASS_UUIDS:
             case BLUETOOTH_DATA_TYPE_COMPLETE_LIST_OF_16_BIT_SERVICE_CLASS_UUIDS:
-                for (i=0; (i + (uint8_t) 2) <= data_len; i+= (uint8_t) 2){
+                for (i=0u; (i + 2u) <= data_len; i+= 2u){
                     uint16_t uuid = (uint16_t) little_endian_read_16(data, (int) i);
                     if ( uuid == uuid16 ) {
                         return true;
@@ -123,10 +123,10 @@ bool ad_data_contains_uuid16(uint8_t ad_len, const uint8_t * ad_data, uint16_t u
             case BLUETOOTH_DATA_TYPE_COMPLETE_LIST_OF_128_BIT_SERVICE_CLASS_UUIDS:
                 uuid_add_bluetooth_prefix(ad_uuid128, uuid16);
                 reverse_128(ad_uuid128, uuid128_bt);
-                for (i=0; (i + (uint8_t) 16) <= data_len; i += (uint8_t) 16){
-                  if (memcmp(uuid128_bt, &data[i], 16) == 0){
-                      return true;
-                  };
+                for (i=0u; (i + 16u) <= data_len; i += 16u){
+                    if (memcmp(uuid128_bt, &data[i], 16) == 0){
+                        return true;
+                    };
                 }
                 break;
             default:
@@ -154,7 +154,7 @@ bool ad_data_contains_uuid128(uint8_t ad_len, const uint8_t * ad_data, const uin
         switch (data_type){
             case BLUETOOTH_DATA_TYPE_INCOMPLETE_LIST_OF_16_BIT_SERVICE_CLASS_UUIDS:
             case BLUETOOTH_DATA_TYPE_COMPLETE_LIST_OF_16_BIT_SERVICE_CLASS_UUIDS:
-                for (i = (uint8_t) 0; (i+ (uint8_t) 2) <= data_len; i += (uint8_t) 2){
+                for (i = 0u; (i+2u) <= data_len; i += 2u){
                     uint16_t uuid16 = little_endian_read_16(data, (int) i);
                     uuid_add_bluetooth_prefix(ad_uuid128, uuid16);
                     
@@ -166,7 +166,7 @@ bool ad_data_contains_uuid128(uint8_t ad_len, const uint8_t * ad_data, const uin
                 break;
             case BLUETOOTH_DATA_TYPE_INCOMPLETE_LIST_OF_128_BIT_SERVICE_CLASS_UUIDS:
             case BLUETOOTH_DATA_TYPE_COMPLETE_LIST_OF_128_BIT_SERVICE_CLASS_UUIDS:
-                for (i = (uint8_t) 0; (i + (uint8_t) 16) <= data_len; i += (uint8_t) 16){
+                for (i = 0u; (i + 16u) <= data_len; i += 16u){
                   if (memcmp(uuid128_le, &data[i], 16) == 0) {
                       return true;
                   }
